Add missing uint64_t type definition and clean up includes in joltage files

+++ include/joltage.hpp
@@ -1,5 +1,6 @@
 #pragma once
 #include <string>
+#include <cstdint>
 
 int max_joltage_from_bank(const std::string &s);
-unsigned long long max12_joltage_from_bank(const std::string &s);
+uint64_t max12_joltage_from_bank(const std::string &s);
